Move cast.

+++ src/SevenEval.h
@@ -46,7 +46,7 @@ public:
         return flush_ranks[s[i] | s[j] | s[k] | s[m] | s[n] | s[p] | s[q]];
       }
     }
-    auto const hash = FACE_BIT_MASK & (uint32_t)(31ULL * key);
+    auto const hash = FACE_BIT_MASK & (uint32_t)(31 * (uint64_t) key);
     return rank_hash[offsets[hash >> RANK_OFFSET_SHIFT] +
       (hash & RANK_HASH_MOD)];
   }
